Set AC_CONFIG_MACRO_DIRS([m4])
As recommended by libtool.  Along with adding the --install option to
aclocal, which causes the third-party macros (from autoconf-archive) to
be copied into the m4 directory rather than appending to aclocal.m4.

This should help with comprehensibility and maintainability of the
build scripts (as much as that is possible with autotools...) and make
it easier to add custom macros later.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>