Support forcing 2-character arguments as short
When ULTRAGETOPT_2CHAR_ARG_SHORT is defined (or UGO_2CHARARGSHORT passed
to ultragetopt_tuneable) any 2-character argument for a valid short
option is interpreted as a short option.

This is the way glib getopt_long_only behaves.  It is not the default
behavior because it prevents single-character long options from ever
matching (while short options would still match as non-first options).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>