Fix up test library checking
Do not add test libraries to LIBS (which gets linked into everything).
Instead, add the test libraries to TEST_LIBS and reference it in LDADD
of test programs.

Note:  Failing to find a test library is non-fatal, since Ultragetopt
may still function on systems where the test libraries do not.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>