BREAKING:  Change default optarg, optopt behavior
Initialize optopt to '?' in all cases (since this is done by both BSD
and GNU getopt).

Rename ULTRAGETOPT_DEFAULTOPTOPT to ULTRAGETOPT_DEFAULT_OPTOPT and set
optopt to this value on every call if defined.

Set optarg to NULL on every call unless ULTRAGETOPT_NO_DEFAULT_OPTARG is
defined, in which case it is not set by getopt but is still set by
getopt_long and getopt_long_only.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>